Seperated time, code cleanup based on comments
I separated the time to the localTime file, created a hook for it and corrected the code with the comments provided in the previous pull request
