+++ src/styled.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-
+//header
 export const Navigation = styled.div`
   @import url("https://fonts.googleapis.com/css2?family=Montserrat&display=swap");
   font-family: "Montserrat", sans-serif;
@@ -32,6 +32,7 @@ export const Image = styled.img`
   height: 100px;
 `;
 
+//input
 export const Input = styled.form`
   display: flex;
   width: 47.5%;
@@ -80,6 +81,7 @@ export const Box = styled.input`
 }
 `;
 
+//singletodo
 export const Single = styled.form`
   display: flex;
   border-radius: 5px;
@@ -121,6 +123,7 @@ export const Icon = styled.span`
   cursor: pointer;
 `;
 
+//todolist
 export const Todos = styled.div`
   border-radius: 5px;
   display: flex;
@@ -129,6 +132,7 @@ export const Todos = styled.div`
   padding: 15px;
 `;
 
+//app
 export const Container = styled.div`
   @import url("https://fonts.googleapis.com/css2?family=Montserrat&display=swap");
   width: 100vw;
